+++ src/core/turn-order.js
@@ -47,20 +47,22 @@ export const Pass = (G, ctx) => {
  *                       // (after which they're pruned from actionPlayers).
  *   }
  */
-export function SetActionPlayers(state, arg) {
+export function SetActionPlayersEvent(state, arg) {
+  return { ...state, ctx: setActionPlayers(state.ctx, arg) };
+}
+
+function setActionPlayers(ctx, arg) {
   let actionPlayers = [];
 
   if (arg.value) {
     actionPlayers = arg.value;
   }
   if (arg.all) {
-    actionPlayers = [...state.ctx.playOrder];
+    actionPlayers = [...ctx.playOrder];
   }
 
   if (arg.allOthers) {
-    actionPlayers = [...state.ctx.playOrder].filter(
-      nr => nr !== state.ctx.currentPlayer
-    );
+    actionPlayers = [...ctx.playOrder].filter(nr => nr !== ctx.currentPlayer);
   }
 
   if (Array.isArray(arg)) {
@@ -68,13 +70,10 @@ export function SetActionPlayers(state, arg) {
   }
 
   return {
-    ...state,
-    ctx: {
-      ...state.ctx,
-      actionPlayers,
-      _actionPlayersOnce: arg.once,
-      _actionPlayersAllOthers: arg.allOthers,
-    },
+    ...ctx,
+    actionPlayers,
+    _actionPlayersOnce: arg.once,
+    _actionPlayersAllOthers: arg.allOthers,
   };
 }
 
@@ -94,26 +93,16 @@ function getCurrentPlayer(playOrder, playOrderPos) {
  * @param {object} turnOrder - A turn order object for this phase.
  */
 export function InitTurnOrderState(G, ctx, turnOrder) {
-  let playOrderPos;
-  let actionPlayers;
-
-  const t = turnOrder.first(G, ctx);
-
-  if (t.playOrderPos !== undefined) {
-    playOrderPos = t.playOrderPos;
-  } else {
-    playOrderPos = t;
-  }
-
+  const playOrderPos = turnOrder.first(G, ctx);
   const currentPlayer = getCurrentPlayer(ctx.playOrder, playOrderPos);
 
-  if (t.actionPlayers !== undefined) {
-    actionPlayers = t.actionPlayers;
+  if (turnOrder.actionPlayers !== undefined) {
+    ctx = setActionPlayers(ctx, turnOrder.actionPlayers);
   } else {
-    actionPlayers = [currentPlayer];
+    ctx = { ...ctx, actionPlayers: [currentPlayer] };
   }
 
-  return { ...ctx, currentPlayer, playOrderPos, actionPlayers };
+  return { ...ctx, currentPlayer, playOrderPos };
 }
 
 /**
@@ -141,20 +130,13 @@ export function UpdateTurnOrderState(G, ctx, turnOrder, endTurnArg) {
   } else {
     const t = turnOrder.next(G, ctx);
 
-    if (t == undefined) {
+    if (t === undefined) {
       endPhase = true;
     } else {
-      if (t.playOrderPos !== undefined) {
-        playOrderPos = t.playOrderPos;
-      } else {
-        playOrderPos = t;
-      }
-
+      playOrderPos = t;
       currentPlayer = getCurrentPlayer(ctx.playOrder, playOrderPos);
 
-      if (t.actionPlayers !== undefined) {
-        actionPlayers = t.actionPlayers;
-      } else {
+      if (turnOrder.actionPlayers === undefined) {
         actionPlayers = [currentPlayer];
       }
     }
@@ -179,9 +161,9 @@ export function UpdateTurnOrderState(G, ctx, turnOrder, endTurnArg) {
  * begins, and also a function `next` to determine who the
  * next player is when the turn ends.
  *
- * first / next can also return an object of type
- * { playOrderPos, actionPlayers }
- * in which case they can also set actionPlayers simultaneously.
+ * Objects can also contain an actionPlayers section which
+ * is passed to SetActionPlayers above at the beginning of
+ * the phase.
  *
  * The phase ends if next() returns undefined.
  */
@@ -217,16 +199,9 @@ export const TurnOrder = {
    * currentPlayer switches around in round-robin fashion, but any player can play on each turn.
    */
   ANY: {
-    first: (G, ctx) => {
-      return {
-        actionPlayers: [...ctx.playOrder],
-        playOrderPos: ctx.playOrderPos,
-      };
-    },
-    next: (G, ctx) => {
-      const playOrderPos = (ctx.playOrderPos + 1) % ctx.playOrder.length;
-      return { actionPlayers: [...ctx.playOrder], playOrderPos };
-    },
+    first: (G, ctx) => ctx.playOrderPos,
+    next: (G, ctx) => (ctx.playOrderPos + 1) % ctx.playOrder.length,
+    actionPlayers: { all: true },
   },
 
   /**